requirements.txt
altair>=6.0.0,<7.0.0
pandas>=2.3.0,<3.0.0
streamlit>=1.52.0,<2.0.0
python-dotenv>=1.2.0,<2.0.0
streamlit-authenticator==0.2.3
azure-storage-blob>=12.28.0,<13.0.0
azure-identity>=1.25.0,<2.0.0
pyyaml>=6.0.0,<7.0.0